- thanks, John for the hint with the unsigned short int!  Reduced   mem requirements for the adjacent_vertices matrices of all higher-order   elements through converting to short int - reduced # of virtual fn's in Elem by one: the higher_order_equiv_elem()   or so got moved into a static member function of Elem.  This also enables   to choose whether one wants to convert quad4->quad8 or quad4->quad9   through a simple bool option in MeshBase::all_second_order(bool) - reduced repeated code in adjacent_vertices through moving these short   int matrices to Hex, Quad, Prism etc - fixed GMV support for higher-order Prisms (enuff debugging for today... ;-)
git-svn-id: file:///Users/petejw/Documents/libmesh_svn_bak@426 434f946d-2f3d-0410-ba4c-cb9f52fb0dbf

+++ trunk/libmesh/include/cell_inf_hex18.h
@@ -1,4 +1,4 @@
-// $Id: cell_inf_hex18.h,v 1.16 2003-08-07 19:25:30 ddreyer Exp $
+// $Id: cell_inf_hex18.h,v 1.17 2003-08-18 14:44:51 ddreyer Exp $
 
 // The Next Great Finite Element Library.
 // Copyright (C) 2002  Benjamin S. Kirk, John W. Peterson
@@ -135,8 +135,8 @@ public:
    * that defines the \f$ n^{th} \f$ second-order node.
    * Note that \p n is counted as depicted above, \f$ 8 \le n < 18 \f$.
    */
-  unsigned int second_order_adjacent_vertex (const unsigned int n,
-					     const unsigned int v) const;
+  unsigned short int second_order_adjacent_vertex (const unsigned int n,
+						   const unsigned int v) const;
    
 
 protected:
